backend/requirements.txt
blinker
certifi
charset-normalizer
click
colorama
urllib3
Werkzeug
Flask
requests
Flask-SQLAlchemy
gunicorn
pymysql
Flask-Migrate
cryptography
mysql-connector-python
python-dotenv
schedule
flask-cors
Flask
bcrypt
